core refactor
