+++ packages/ultrahonk/src/decider/sumcheck/verifier.rs
@@ -14,19 +14,38 @@ use crate::{
     CONST_PROOF_SIZE_LOG_N,
 };
 use alloc::vec::Vec;
-use ark_ff::Zero;
+use ark_ff::{One, Zero};
+use crate::decider::types::BATCHED_RELATION_PARTIAL_LENGTH;
+
 
 // Keep in mind, the UltraHonk protocol (UltraFlavor) does not per default have ZK
 impl<P: G1ArithmeticBackend, H: HashBackend> DeciderVerifier<P, H> {
-    pub(crate) fn sumcheck_verify<const SIZE: usize>(
+    pub fn verify_sumcheck(
         &mut self,
         transcript: &mut Transcript,
-        padding_indicator_array: &[ScalarField; CONST_PROOF_SIZE_LOG_N],
+        circuit_size: u32,
     ) -> HonkVerifyResult<SumcheckVerifierOutput> {
+        let log_circuit_size = circuit_size.ilog2();
+
+        // determine padding indicator array
+        let mut padding_indicator_array = [ScalarField::zero(); CONST_PROOF_SIZE_LOG_N];
+        for (idx, value) in padding_indicator_array.iter_mut().enumerate() {
+            *value = if idx < log_circuit_size as usize {
+                ScalarField::one()
+            } else {
+                ScalarField::zero()
+            };
+        }
+
         let mut verified: bool = true;
 
         // Pad gate challenges for Protogalaxy DeciderVerifier and AVM
-        self.pad_gate_challenges();
+        if self.memory.relation_parameters.gate_challenges.len() < CONST_PROOF_SIZE_LOG_N {
+            let zero = ScalarField::zero();
+            for _ in self.memory.relation_parameters.gate_challenges.len()..CONST_PROOF_SIZE_LOG_N {
+                self.memory.relation_parameters.gate_challenges.push(zero);
+            }
+        }
 
         let mut gate_separators = GateSeparatorPolynomial::new_without_products(
             self.memory.relation_parameters.gate_challenges.to_owned(),
@@ -39,7 +58,7 @@ impl<P: G1ArithmeticBackend, H: HashBackend> DeciderVerifier<P, H> {
             let round_univariate_label = format!("Sumcheck:univariate_{}", round_idx);
 
             let evaluations =
-                transcript.receive_fr_array_from_verifier::<SIZE>(round_univariate_label)?;
+                transcript.receive_fr_array_from_verifier::<BATCHED_RELATION_PARTIAL_LENGTH>(round_univariate_label)?;
             let round_univariate = SumcheckRoundOutput { evaluations };
 
             let round_challenge = transcript.get_challenge::<H>(format!("Sumcheck:u_{}", round_idx));
@@ -74,7 +93,6 @@ impl<P: G1ArithmeticBackend, H: HashBackend> DeciderVerifier<P, H> {
         }
 
         // Evaluate the Honk relation at the point (u_0, ..., u_{d-1}) using claimed evaluations of prover polynomials.
-
         let full_honk_purported_value =
             SumcheckVerifierRound::compute_full_relation_purported_value(
                 &self.memory.claimed_evaluations,
@@ -88,14 +106,5 @@ impl<P: G1ArithmeticBackend, H: HashBackend> DeciderVerifier<P, H> {
             multivariate_challenge,
             verified,
         })
-    }
-
-    fn pad_gate_challenges(&mut self) {
-        if self.memory.relation_parameters.gate_challenges.len() < CONST_PROOF_SIZE_LOG_N {
-            let zero = ScalarField::zero();
-            for _ in self.memory.relation_parameters.gate_challenges.len()..CONST_PROOF_SIZE_LOG_N {
-                self.memory.relation_parameters.gate_challenges.push(zero);
-            }
-        }
     }
 }
